AssetGroupManager with git_remote options
